implemented tiling script for tileview.. more work on tileview app version

+++ Tensorflow/prepare_for_tablet.py
@@ -19,7 +19,16 @@ import json
 import os
 import pyproj
 import progressbar
+import math
+from utils import file_utils
+import numpy as np
+from PIL import Image
+from concurrent.futures import ThreadPoolExecutor
+import time
+import signal                         
 
+
+stop = False
 
 class GeoInformation(object):
     def __init__(self):
@@ -29,7 +38,7 @@ class GeoInformation(object):
         self.ul_lat = 0
      
     
-def preprocess_internal(in_path, out_path, tile_size = 5120):
+def preprocess_internal(in_path, out_path, tile_size = 256):
     """
     Given an input-image (any format) and an output-folder,
     the command tiles the input-image into tiles of suitable size for an android
@@ -48,12 +57,6 @@ def preprocess_internal(in_path, out_path, tile_size = 5120):
     Returns:
         None, a folder that can bi copied onto the android tablet.
     """
-
-    #fixed tile size
-    output_filename = 'tile_'
-    tile_size_x = tile_size
-    tile_size_y = tile_size
-    
     #Read the input Coordinate system
     ds = gdal.Open(in_path)
     inSRS_wkt = ds.GetProjection()  # gives SRS in WKT
@@ -81,40 +84,84 @@ def preprocess_internal(in_path, out_path, tile_size = 5120):
     xsize = band.XSize
     ysize = band.YSize
     
-    print("Tiling image...")
-    #Tiling the image
-    for i in progressbar.progressbar(range(0, xsize, tile_size_x)):
-        for j in range(0, ysize, tile_size_y):
-            
-            #define paths of image tile and the corresponding json file containing the geo information
-            out_path_image = str(out_path) + str(output_filename) + "row" + str(int(j/tile_size_y)) + "_col" + str(int(i/tile_size_x)) + ".png"
-            out_path_json = str(out_path) + str(output_filename) + "row" + str(int(j/tile_size_y)) + "_col" + str(int(i/tile_size_x)) + "_geoinfo.json"
-            
-            #tile image with gdal (copy bands 1, 2 and 3)
-            gdal.Translate(out_path_image,ds, options=gdal.TranslateOptions(srcWin=[i,j,tile_size_x,tile_size_y], bandList=[1,2,3], format='png'))
-            
-
-            if create_geoinfo_file:
-                #calculate upper left and lower right coordinates of image tile
-                geo_info_curr = GeoInformation()
-                geo_info_curr.ul_lon = ulx + (i * xres)
-                geo_info_curr.ul_lat = uly + j* yres
-                geo_info_curr.lr_lon = ulx + (i+tile_size_x) * xres
-                geo_info_curr.lr_lat = uly + (j+tile_size_x) * yres
-                
-                #convert them to wgs84
-                geo_info_curr.lr_lon,geo_info_curr.lr_lat  = pyproj.transform(swiss, wgs84, geo_info_curr.lr_lon, geo_info_curr.lr_lat)
-                geo_info_curr.ul_lon,geo_info_curr.ul_lat = pyproj.transform(swiss, wgs84, geo_info_curr.ul_lon, geo_info_curr.ul_lat)
-
-                #save geoinfo file
-                with open(out_path_json, 'w') as outfile:
-                    json.dump(geo_info_curr.__dict__, outfile)
+    print("Loading input image of size " + str(xsize) + "x" + str(ysize) + "...")
+    image_array = ds.ReadAsArray().astype(np.uint8)
+    image_array = np.swapaxes(image_array,0,1)
+    image_array = np.swapaxes(image_array,1,2)
     
+    
+    
+    print("Tiling image...")
+    
+    futures_list = []
+    with ThreadPoolExecutor(max_workers=20) as executor:
+
+    
+    
+        number_of_zoom_levels = int(min(math.floor(math.log(xsize/tile_size,2)), math.floor(math.log(ysize/tile_size,2))))
+        
+        for zoom_level in (range(0,number_of_zoom_levels)):
+            
+            os.makedirs(os.path.join(out_path,str(zoom_level)),exist_ok=True)
+            original_tilesize = tile_size*2**(zoom_level)
+            #Tiling the image
+            for x_start in range(0, xsize, original_tilesize):
+                for y_start in range(0, ysize, original_tilesize):
+                    
+                    #define paths of image tile and the corresponding json file containing the geo information
+                    out_path_image = str(out_path) + str(zoom_level) + "/tile_level" + str(zoom_level) + "_x" + str(int(x_start/original_tilesize)) + "_y" + str(int(y_start/original_tilesize)) + ".png"
+                    #out_path_json = str(out_path) + str(output_filename) + "row" + str(int(j/tile_size_y)) + "_col" + str(int(i/tile_size_x)) + "_geoinfo.json"
+
+                    future = executor.submit(crop_image,ds,tile_size, original_tilesize,image_array,y_start,x_start,out_path_image)
+                    futures_list.append(future)
+                    
+        for i in progressbar.progressbar(range(0,len(futures_list))):
+            while not futures_list[i].done():
+                time.sleep(0.5)
+                   
+
+
+                
+        '''
+        if create_geoinfo_file:
+            #calculate upper left and lower right coordinates of image tile
+            geo_info_curr = GeoInformation()
+            geo_info_curr.ul_lon = ulx + (i * xres)
+            geo_info_curr.ul_lat = uly + j* yres
+            geo_info_curr.lr_lon = ulx + (i+tile_size_x) * xres
+            geo_info_curr.lr_lat = uly + (j+tile_size_x) * yres
+            
+            #convert them to wgs84
+            geo_info_curr.lr_lon,geo_info_curr.lr_lat  = pyproj.transform(swiss, wgs84, geo_info_curr.lr_lon, geo_info_curr.lr_lat)
+            geo_info_curr.ul_lon,geo_info_curr.ul_lat = pyproj.transform(swiss, wgs84, geo_info_curr.ul_lon, geo_info_curr.ul_lat)
+
+            #save geoinfo file
+            with open(out_path_json, 'w') as outfile:
+                json.dump(geo_info_curr.__dict__, outfile)
+        '''
+
+
+
+def crop_image(tile_size, ds, original_tilesize,image_array,y_start,x_start,out_path_image):
+    
+    if not stop:
+        if original_tilesize < 10000:
+                        #crop_image(tile_size, original_tilesize,image_array,y_start,x_start,out_path_image)
+            cropped_array = image_array[y_start:y_start+original_tilesize,x_start:x_start+original_tilesize,:]
+            
+            pad_end_x = original_tilesize - cropped_array.shape[1]
+            pad_end_y = original_tilesize - cropped_array.shape[0]
+            cropped_array = np.pad(cropped_array,((0,pad_end_y),(0,pad_end_x),(0,0)), mode='constant', constant_values=0)
+            cropped_im = Image.fromarray(cropped_array,'RGB')
+            cropped_im = cropped_im.resize((tile_size,tile_size), Image.ANTIALIAS)
+            #cropped_im = cropped_im.convert("RGB")
+            cropped_im.save(out_path_image)
+        else:
+            gdal.Translate(out_path_image,ds, options=gdal.TranslateOptions(width=int(tile_size),height=int(tile_size),srcWin=[x_start,y_start,original_tilesize,original_tilesize], bandList=[1,2,3], format='png'))
             if os.path.isfile(out_path_image + ".aux.xml"):
                 os.remove(out_path_image + ".aux.xml")
 
-        
-    
+          
 def preprocess(in_path, out_path,tile_size=5120):
     """
     Does some sanity checks on the input parameters before passing them on to
@@ -151,7 +198,15 @@ def preprocess(in_path, out_path,tile_size=5120):
         print("Input File must be a .tif file!")
         return
     '''
+    def signal_handling(signum,frame):           
+        global stop                         
+        stop = True  
+        print("\nstopping...")     
+
+    signal.signal(signal.SIGINT,signal_handling) 
+
     preprocess_internal(in_path,out_path,tile_size)
+
